fix errors about pin memory for regconition model

+++ trainer/craft/train.sh
@@ -1,12 +1,18 @@
 # Create conda environment and install dependencies
 base_env=$(conda info | grep -i 'base environment' | awk -F': ' '{print $2}' | sed 's/ (read only)//' | tr -d ' ')
-conda create --name easyocr_detection python=3.8 -y
+env_name=easyocr_detection
 source ${base_env}/etc/profile.d/conda.sh
-conda activate easyocr_detection 
-pip install -r requirements.txt
-moreh-switch-model -M 2
-echo -e "\\n" | update-moreh --torch 1.13.1 --target 23.6.0 --force
 
+if conda env list | grep -q -E "^$env_name\s"; then
+    conda activate ${env_name}
+else
+    conda clean --all --force-pkgs-dir -y
+    conda create --name ${env_name} python=3.8 -y
+    conda activate ${env_name}
+    pip install -r requirements.txt
+    moreh-switch-model -M 2
+    echo -e "\\n" | update-moreh --torch 1.13.1 --target 23.6.0 --force
+fi
 # Download pretrained model
 model_dir="/nas/thuchk/repos/EasyOCR/trainer/craft/pretrained_model/"
 mkdir -p "${model_dir}"
@@ -25,4 +31,4 @@ CUDA_VISIBLE_DEVICES=0 python3 train.py --yaml=custom_data_train
 # Delete conda environment
 echo "deleting env.."
 conda deactivate
-conda env remove -n easyocr_detection
+conda env remove -n $env_name
